+++ requake/plot/plot_cumulative.py
@@ -19,7 +19,7 @@ from ..families import FamilyNotFoundError, read_selected_families
 from ..formulas import mag_to_slip_in_cm, mag_to_moment
 from .plot_utils import (
     format_time_axis, plot_title, hover_annotation, duration_string,
-    family_colors, plot_colorbar
+    family_colors, plot_colorbar, save_or_show_plot
 )
 logger = logging.getLogger(__name__.rsplit('.', maxsplit=1)[-1])
 # Reduce logging level for Matplotlib to avoid DEBUG messages
@@ -174,4 +174,4 @@ def plot_cumulative():
     annot.set_visible(False)
     annot.hover_annotation = True
     fig.canvas.mpl_connect('motion_notify_event', hover_annotation)
-    plt.show()
+    save_or_show_plot(fig, 'family_cumulative')